react-flask-build/flask-server/requirements.txt
Flask
Flask-CORS
python-dotenv
tiktoken
openai
torch
transformers
PyMuPDF
